refactor: API 폴더로 분리
