+++ src/domain/MovieClient.ts
@@ -1,8 +1,8 @@
-import APIClient from './APIClient';
+import APIClient from './API/APIClient';
 import { SYSTEM_CONSTANTS } from '../constants/systemConstants';
 import { Movie } from '../../types/apiMovie';
 import { redirectToPage } from '../route/router';
-import { convertToMovieData, convertToMovieDetailData } from './converToData';
+import { convertToMovieData, convertToMovieDetailData } from './API/converToData';
 
 class MovieClient extends APIClient {
   constructor() {
